requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0